% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{Dynamic Programming Knapsack}
\source{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{A dataframe containing objects to be put in the knapsack, must contain two variables: v and w}

\item{W}{Knapsack size}
}
\description{
Dynamic Programming Knapsack
}
