% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Algorithm Knapsack}
\source{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A dataframe containing objects to be put in the knapsack, must contain two variables: v and w}

\item{W}{Knapsack size}
}
\description{
Greedy Algorithm Knapsack
}
