% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_knapsack.R
\docType{data}
\name{test_knapsack}
\alias{test_knapsack}
\title{A data set containing a graph taken from the Wikipedia page of Dijkstra's algorithm.}
\format{
A data frame with knapsack elements for testing
\describe{
  \item{v}{Items values}
  \item{w}{Items weights}
  }
}
\source{
Lab 6 Instructions
}
\usage{
test_knapsack
}
\description{
A data set containing a graph taken from the Wikipedia page of Dijkstra's algorithm.
}
\keyword{datasets}
